Leaderboard functionality
Display function added to display JFrame.
GetMax and GetMin have been updated to handle when the list is empty.
Remove score will always remove the last score instead of the hard coded index of 9.
Add score will now check if list is empty, if they have a score less than min but there are not 10 scores yet, and finally a binary search to find the placement if the list is full.